requirements.txt
alembic==1.6.5
feedparser==6.0.8
pandas==1.3.1
PyYAML==5.1.2
pathlib2==2.3.4
SQLAlchemy==1.3.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
psycopg2==2.9.3
psycopg2-binary==2.9.3